+++ src/ManagementPortal/Api/Views/Applications/CreateNewApplication.cshtml
@@ -1,4 +1,48 @@
-@model ManagementPortal.Api.Views.Applications.CreateNewApplicationModel
+@using ManagementPortal.Api.Views.Applications.Shared
+@model CreateNewApplicationModel
 
+@{
+    Layout = "_Layout";
+    ViewData["Title"] = "Create New Client";
+    
+    // Map the model to the shared form model
+    var formModel = new ApplicationFormModel
+    {
+        IsEditMode = false,
+        Id = Model.Id,
+        ApplicationType = Model.ApplicationType,
+        ClientId = Model.ClientId,
+        ClientType = Model.ClientType,
+        ConsentType = Model.ConsentType,
+        DisplayName = Model.DisplayName,
+        DisplayNames = Model.DisplayNames,
+        JsonWebKeySet = Model.JsonWebKeySet,
+        Permissions = Model.Permissions,
+        PostLogoutRedirectUris = Model.PostLogoutRedirectUris,
+        Properties = Model.Properties,
+        RedirectUris = Model.RedirectUris,
+        Requirements = Model.Requirements,
+        Settings = Model.Settings
+    };
+}
 
+<div class="bg-white shadow overflow-hidden sm:rounded-lg">
+    <div class="flex justify-between items-center px-6 py-4 border-b border-gray-200">
+        <h1 class="text-2xl font-semibold text-gray-900">Create New OIDC Client</h1>
+        <a asp-action="Index"
+           class="inline-flex items-center px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-gray-600 hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-500">
+            <svg class="-ml-1 mr-2 h-5 w-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                <path fill-rule="evenodd"
+                      d="M9.707 16.707a1 1 0 01-1.414 0l-6-6a1 1 0 010-1.414l6-6a1 1 0 011.414 1.414L5.414 9H17a1 1 0 110 2H5.414l4.293 4.293a1 1 0 010 1.414z"
+                      clip-rule="evenodd"/>
+            </svg>
+            Back to List
+        </a>
+    </div>
 
+    <div class="px-6 py-5">
+        <form method="post">
+            <partial name="Shared/_ApplicationForm" model="@formModel" />
+        </form>
+    </div>
+</div>
