+++ src/ManagementPortal/Api/Views/Applications/Shared/_DynamicListInput.cshtml
@@ -1,4 +1,4 @@
-@model DynamicListInput
+@model _DynamicListInput
 @inject Microsoft.AspNetCore.Antiforgery.IAntiforgery Antiforgery
 
 <div>
