+++ src/ManagementPortal/Api/Views/Applications/Shared/_DynamicListItem.cshtml
@@ -1,4 +1,4 @@
-@model ManagementPortal.Api.Views.Applications.Edit.DynamicListItem
+@model _DynamicListItem
 @inject Microsoft.AspNetCore.Antiforgery.IAntiforgery Antiforgery
 
 <div class="redirect-uri-item flex items-center space-x-2 p-2 border border-gray-200 rounded-md bg-gray-50 hx-url-item"
