+++ src/ManagementPortal/Api/Views/Shared/_Layout.cshtml
@@ -17,12 +17,12 @@
                     <span class="text-white text-xl font-bold">OIDC Management</span>
                 </div>
                 <div class="hidden md:ml-6 md:flex md:items-center md:space-x-4">
-                    <a href="/"
+                    <a asp-controller="MainPage" asp-action="Index"
                        class="text-white hover:bg-indigo-500 hover:text-white px-3 py-2 rounded-md text-sm font-medium">Dashboard</a>
                     <a href="/users"
                        class="text-white hover:bg-indigo-500 hover:text-white px-3 py-2 rounded-md text-sm font-medium">Users</a>
-                    <a href="/clients"
-                       class="text-white hover:bg-indigo-500 hover:text-white px-3 py-2 rounded-md text-sm font-medium">Clients</a>
+                    <a  asp-controller="Applications" asp-action="Index"
+                       class="text-white hover:bg-indigo-500 hover:text-white px-3 py-2 rounded-md text-sm font-medium">Applications</a>
                 </div>
             </div>
             <div class="flex items-center">
